expense_sharing/requirements.txt
Django==5.1.2
djangorestframework==3.15.2
django-cors-headers==4.5.0
djangorestframework-simplejwt==5.3.1
reportlab==4.2.5